+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGE LOG
 
-## [Unreleased]
+## [6.5.1]
 
 ### Fixed